requirements.txt
annotate==2.1.1
jsonlines==3.0.0
numpy==1.20.3
pandas==2.1.0
pillow==10.0.0
python-dotenv==0.19.2
requests==2.31.0
streamlit==1.27.0
